@@ -21,17 +21,63 @@ public partial class SurfTimer
         if (bot_quota != null)
         {
             int cbq = bot_quota.GetPrimitiveValue<int>();
-            if(cbq != CurrentMap.ReplayBots.Count)
+
+            int replaybot_count = 1 +
+                                (CurrentMap.ReplayManager.StageWR != null ? 1 : 0) +
+                                (CurrentMap.ReplayManager.BonusWR != null ? 1 : 0) +
+                                CurrentMap.ReplayManager.CustomReplays.Count;
+
+            if (cbq != replaybot_count)
             {
-                bot_quota.SetValue(CurrentMap.ReplayBots.Count);
+                bot_quota.SetValue(replaybot_count);
             }
         }
 
-        for(int i = 0; i < CurrentMap!.ReplayBots.Count; i++)
+        // Console.WriteLine($"======== public void OnTick -> Got bot_quota {cbq} | Setting to bot_quota {replaybot_count}");
+        CurrentMap.ReplayManager.MapWR.Tick();
+        CurrentMap.ReplayManager.StageWR?.Tick();
+        CurrentMap.ReplayManager.BonusWR?.Tick();
+
+        // Here we will load the NEXT stage replay from AllStageWR
+        if (CurrentMap.ReplayManager.StageWR?.RepeatCount == 0)
         {
-            CurrentMap.ReplayBots[i].Tick();
-            if (CurrentMap.ReplayBots[i].RepeatCount == 0)
+            int next_stage;
+            if (CurrentMap.ReplayManager.AllStageWR[(CurrentMap.ReplayManager.StageWR.Stage % CurrentMap.Stages) + 1][0].MapTimeID == -1)
+                next_stage = 1;
+            else
+                next_stage = (CurrentMap.ReplayManager.StageWR.Stage % CurrentMap.Stages) + 1;
+
+            CurrentMap.ReplayManager.AllStageWR[next_stage][0].Controller = CurrentMap.ReplayManager.StageWR.Controller;
+
+            // Console.WriteLine($"======== public void OnTick() -> Finished replay cycle for stage {CurrentMap.ReplayManager.StageWR.Stage}, changing to stage {next_stage}");
+            CurrentMap.ReplayManager.StageWR = CurrentMap.ReplayManager.AllStageWR[next_stage][0];
+            CurrentMap.ReplayManager.StageWR.LoadReplayData(repeat_count: 3);
+            CurrentMap.ReplayManager.StageWR.FormatBotName();
+            CurrentMap.ReplayManager.StageWR.Start();
+        }
+
+        if (CurrentMap.ReplayManager.BonusWR?.RepeatCount == 0)
+        {
+            CurrentMap.ReplayManager.BonusWR.Stage = (CurrentMap.ReplayManager.BonusWR.Stage % CurrentMap.Bonuses) + 1;
+            CurrentMap.ReplayManager.BonusWR.LoadReplayData(repeat_count: 3);
+
+            AddTimer(1.5f, () => CurrentMap.ReplayManager.BonusWR.FormatBotName());
+
+            // CurrentMap.ReplayManager.BonusWR.ResetReplay();
+            // //CurrentMap.ReplayManager.BonusWR.FormatBotName();
+            // CurrentMap.ReplayManager.BonusWR.RepeatCount = 3;
+        }
+
+        for (int i = 0; i < CurrentMap.ReplayManager.CustomReplays.Count; i++)
+        {
+            if (CurrentMap.ReplayManager.CustomReplays[i].MapID != CurrentMap.ID)
+                CurrentMap.ReplayManager.CustomReplays[i].MapID = CurrentMap.ID;
+
+            CurrentMap.ReplayManager.CustomReplays[i].Tick();
+            if (CurrentMap.ReplayManager.CustomReplays[i].RepeatCount == 0)
+            {
                 CurrentMap.KickReplayBot(i);
+            }
         }
     }
 }
