@@ -10,18 +10,19 @@ internal class PlayerTimer
     // Modes
     public bool IsPracticeMode { get; set; } = false; // Practice mode toggle
     public bool IsStageMode { get; set; } = false; // Stage mode toggle
+    public bool IsBonusMode { get; set; } = false; // Bonus mode toggle
 
     // Tracking
     public int Stage { get; set; } = 0; // Current stage tracker
     public int Checkpoint {get; set;} = 0; // Current checkpoint tracker
-    public CurrentRun CurrentRunData { get; set; } = new CurrentRun(); // Current RUN data tracker
+    // public CurrentRun CurrentRunData { get; set; } = new CurrentRun(); // Current RUN data tracker
     public int Bonus { get; set; } = 0; // To-do: bonus implementation - Current bonus tracker 
     public int Style { get; set; } = 0; // To-do: functionality for player to change this value and the actual styles implementation - Current style tracker
 
     // Timing
     public int Ticks { get; set; } = 0; // To-do: sub-tick counting? This currently goes on OnTick, which is not sub-tick I believe? Needs investigating
 
-    // Time Formatting
+    // Time Formatting - To-do: Move to player settings maybe?
     public enum TimeFormatStyle
     {
         Compact,
@@ -39,7 +40,8 @@ internal class PlayerTimer
         this.IsPaused = false;
         this.IsPracticeMode = false;
         this.IsStageMode = false;
-        this.CurrentRunData.Reset();
+        this.IsBonusMode = false;
+        // this.CurrentRunData.Reset();
     }
 
     public void Pause()
