@@ -1,24 +1,45 @@
 namespace SurfTimer;
 using CounterStrikeSharp.API.Modules.Utils;
-using CounterStrikeSharp.API.Core;
 
 internal enum ReplayFrameSituation
 {
         NONE,
-        START_RUN,
-        END_RUN,
-        TOUCH_CHECKPOINT,
-        START_STAGE,
-        END_STAGE
+
+        STAGE_ZONE_ENTER,
+        STAGE_ZONE_EXIT,
+
+        START_ZONE_ENTER,
+        START_ZONE_EXIT,
+
+        END_ZONE_ENTER,
+        END_ZONE_EXIT,
+
+        CHECKPOINT_ZONE_ENTER,
+        CHECKPOINT_ZONE_EXIT,
+
+        // START_RUN,
+        // END_RUN,
+        // TOUCH_CHECKPOINT,
+        // START_STAGE,
+        // END_STAGE,
+        // ENTER_STAGE,
 }
 
 [Serializable]
-internal class ReplayFrame 
+internal class ReplayFrame
 {
-        public Vector Pos { get; set; } = new Vector(0, 0, 0);
-        public QAngle Ang { get; set; } = new QAngle(0, 0, 0);
-        public uint Situation { get; set; } = (uint)ReplayFrameSituation.NONE;
-        public ulong Button { get; set; }
+        public float[] pos { get; set; } = { 0, 0, 0 };
+        public float[] ang { get; set; } = { 0, 0, 0 };
+        public ReplayFrameSituation Situation { get; set; } = ReplayFrameSituation.NONE;
         public uint Flags { get; set; }
-        public MoveType_t MoveType { get; set; }
+
+        public Vector GetPos()
+        {
+                return new Vector(this.pos[0], this.pos[1], this.pos[2]);
+        }
+
+        public QAngle GetAng()
+        {
+                return new QAngle(this.ang[0], this.ang[1], this.ang[2]);
+        }
 }
